// src/api/generated/models/getAllOfficesApiV1OfficesGetParams.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Community App V2
 * OpenAPI spec version: 0.1.0
 */
import type { LifecycleStatusFilter } from './lifecycleStatusFilter'
import type { OfficeSortField } from './officeSortField'
import type { SortOrder } from './sortOrder'

export type GetAllOfficesApiV1OfficesGetParams = {
  status?: LifecycleStatusFilter
  sort_by?: OfficeSortField
  order?: SortOrder
  /**
   * Bounding Box: minLon,minLat,maxLon,maxLat
   */
  bbox?: string | null
  /**
   * @minimum 1
   */
  page?: number
  /**
   * @minimum 1
   * @maximum 100
   */
  size?: number
  q?: string | null
}