// src/api/generated/models/officeSortField.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Community App V2
 * OpenAPI spec version: 0.1.0
 */

/**
 * Define supported sort fields for office list queries.
 */
export type OfficeSortField =
  (typeof OfficeSortField)[keyof typeof OfficeSortField]

export const OfficeSortField = {
  created_at: 'created_at',
  name: 'name',
  contact_email: 'contact_email',
} as const